refactor : 사다리게임 step2 전체경로 변경에 따른 리팩토링 작성코드를 step별로 관리하고 싶어서 step2이하로 전체 소스를 이관.
