+++ src/main/java/ladderGame/step2/Main.java
@@ -1,0 +1,10 @@
+package ladderGame.step2;
+
+import ladderGame.step2.controller.LadderController;
+
+public class Main {
+
+  public static void main(String[] args) {
+    LadderController.starGame();
+  }
+}
