+++ src/main/java/ladderGame/step2/controller/LadderController.java
@@ -1,16 +1,15 @@
-package ladderGame.controller;
+package ladderGame.step2.controller;
 
-import ladderGame.service.GamePlay;
-import ladderGame.view.InputView;
-import ladderGame.view.PrintView;
+import ladderGame.step2.service.GamePlay;
+import ladderGame.step2.view.InputView;
+import ladderGame.step2.view.PrintView;
 
 public class LadderController {
 
   private static final String MSG_INSERT_PLAYER_NAMES = "참여할 사람 이름을 입력하세요. (이름은 쉼표(,)로 구분하세요)";
   private static final String MSG_INSERT_HEIGHT = "최대 사다리 높이는 몇 개인가요?";
 
-  public static void main(String[] args) {
-
+  public static void starGame(){
     String stringNames = InputView.inputUserNamesWithMessage(MSG_INSERT_PLAYER_NAMES);
     int height = InputView.inputLadderHeightWithMessage(MSG_INSERT_HEIGHT);
 
@@ -19,4 +18,5 @@ public class LadderController {
     PrintView.printUsersName(gamePlay.getPlayer());
     PrintView.createLadderView(gamePlay.getLines());
   }
+
 }
