+++ src/main/java/ladderGame/step2/model/Line.java
@@ -1,9 +1,9 @@
-package ladderGame.model;
+package ladderGame.step2.model;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
-import ladderGame.service.LineGenerator;
+import ladderGame.step2.service.LineGenerator;
 
 public class Line {
