+++ src/main/java/ladderGame/step2/service/LineGenerator.java
@@ -1,4 +1,4 @@
-package ladderGame.service;
+package ladderGame.step2.service;
 
 import java.util.ArrayList;
 import java.util.List;
