+++ src/test/java/ladderGame/step2/model/PlayersTest.java
@@ -1,14 +1,14 @@
-package ladderGame.model;
+package ladderGame.step2.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import ladderGame.service.CreatePlayer;
+import ladderGame.step2.service.CreatePlayer;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class PlayersTest {
 
-  @DisplayName("플래이어수와 사다리높이를 입력해서 사다리 생성.")
+  @DisplayName("사용자명을 쉼표단위로 여러개 입력하여 사용자를 생성.")
   @Test
   void createPlayersTest() {
     Players players = new Players(CreatePlayer.createUser("user1,user2"));
